Add an ability to categorize tests using attributes
With this change, test classes or individual tests can be categorized
using tests/util.py/set_attr decorator. Categories can be set with
additional information @set_attr(category=3.10) or directly
@set_attr(slow). After categorizing tests, selected tests
can be run using nosetests. For example, tests without category or
test with the category value less than or equal to 3.8 can be selected
with the -A option of the nosetests  using Python-like syntax:
nosetests -A 'not category or category <= 3.8'
When no attribute option is provided to nosetests with -a or -A,
all tests will be selected automatically.