Use certificates from the tests jar and add an extra test
Added a test where SSL is enabled on the client side but
no cafile is set and member is configured with a self-signed certificate.
Also, to use the Let's Encrypt certificate from the enterprise
tests jar, some modifications made to the run-tests and start-rc
scripts. Now, they are also donwloading the enterprise tests jar.
Also updated the PowerShell script for the Winodws.