Fix heartbeat test that fails on Windows
Heartbeat test was failing on windows because
while connecting to the second member that
was created recently, member was throwing
HazelcastInstanceNotActive exception during
the client authentication process. With this fix,
client will try to connect to the second member
and if an exception occurs, it will sleep
for a while and try again up to 3 times.