Fix PagingPredicate#reset and its tests
`list.clear()` is not available on Python2.7. We have missed that
since `reset()` method is not tested. According to the standard library
documentation, `list.clear()` is equivalent to `del list[:]`, so I replaced
it with that.

Also, added some missing tests and replaced `six.assertCountEquals`
with `assertEquals` to check ordering of elements along with equality.