Optimize murmur hash calculation
We were always passing a bytes-like object to the murmur hash function
but, we were copying it from the HEAP_DATA_OVERHEAD(8) offset, which
can be huge. I removed that copy and tuned the offsets according to our
HEAP_DATA_OVERHEAD.

Another parameter of the function was the data size. For valid Data
objects, this is always equal to the `len(buffer) - HEAP_DATA_OVERHEAD`.
This parameter is removed and this length is calculated inside the
murmur hash function now.

Also, inlined the _fmix function call and removed the unnecessary
method calls on hash_to_index function.