Python2 related fixes
Fix the `__hash__` method of the `Data` class to convert
`bytes` to `bytearray` if it executing on Python2 since we cannot
use murmur hash on `str`(alias of `bytes` on Python2) objects. This
extra if check is not necessary on `hash_code` method since
it is only called on the Data objects to be sent by the client
which will have a bytearray buffer.

Also, extra test cases are added