Return early from the handle_write if no data is written
It may be the case that client is disconnected while we are in the
loop, an no data is sent. In that case, we were adding the data
to the write queue again, and try to pop from it within the same loop.
That was causing an infinite loop. An early return is added to
solve this problem. When disconnected, dispatcher will be closed,
and handle_write won't be called again by the asyncore.