Serialization service improvements (#244)
* Optimize the hot path for the serialization of integer values

For integers, unless the user configures the default int type,
it will be 4 bytes integer. Therefore, we first check the default
setting while looking for serailizer for integer types.

* Optimize ObjectDataOutput#to_byte_array

It is not necessary to create a new buffer on `to_byte_array` method
before returning. `buffer[:pos]` already creates a copy.

* Make use of the fact that Data's buffer is never None

There were checks that assume Data's buffer might be None
but this is not possible. Data might have a buffer of length
0, but it can never be None.

This commit removes such checks.

* Make use of the fact that DataOutput's buffer is never None

We initialize the buffer at the constructor and never set it to
None on any of the code paths. Therefore, None checks are not
necessary.

* Determine type id at the serialization service constructor

There is no need to check for int type id on each lookup_default_serializer
call. However, this is required for VAR size option.