Avoid possible race on BackupAcksTest (#254)
It might be the case that the replica node could not finish its
partition table initialization when it receives the backup operation
request from the primary node. That can cause the backup request to
be dropped which can result in test failure since there won't be any
backup acks from the replica node.

This won't be needed when the https://github.com/hazelcast/hazelcast/issues/9398
is fixed. Until then, a call like `map.get(1)` should be enough as a
workaround.

See Also: https://github.com/hazelcast/hazelcast-nodejs-client/pull/693