Make the docs compatible with RTD and update README (#266)
* Make the docs compatible with RTD and update README

It makes more sense to host API documentation and the long README
we have in Read the Docs so that they become more visible and
structured. Therefore, the long README is partitioned into chapters,
a small introduction(docs/index.rst) is created, and API documentation
is moved into a separate chapter. Also, API documentation of the
HazelcastClient is constructed as a top-level chapter since it is
generally the entry point for other documentation elements.

Finally, the README we have in Github is made shorter and
more descriptive. Also, added some badges to both README and
docs/index.rst to make users easily navigate between PyPI, Read the Docs,
and Github. The links to Read the Docs site (hazelcast.readthedocs.io)
won't work right now, but I will set up this once this is merged.

* address review comments