Use the same correlation id on invocation retries
We were using a new correlation id on invocation retries, which
can cause problems because the invocation is modified in-place.

We may have pending invocation entries with the key of correlation
id of X while the correlation id we would get from the client
message of the invocation is Y (Y > X).

It is changed so that the invocation is retried but now uses the
same correlation id.