Add ability to use enum literals in config
This PR adds the ability to use enum literals in the config
to make it easier to set these options, without importing anything.

When a value for enum is set, we first check if it is an actual
value for an enum(i.e., 0 for SSLProtocol.SSLv2). If it is, we
directly use that. If not, and the given configuration value is
a string, we try to get the enum member associated with that name.
If it matches to any member(i.e., SSLv2 for SSLProtocol.SSLv2),
we get the value associated with it and set it as the value.

If none of the above works, we throw TypeError as before.