Initial integration for black
This contains the necessary changes to start using black
with the codebase.

Right now, only the files under the examples and docs folders and
hazelcast/__init__.py is reformatted and checked.

Since black only works on py3.6+, we will only be using it in python3 pr
builder.

Also, added a missing exclude package(`docs.*`) to setup.py and separated
test and dev dependencies.