Add more logging for flaky add_distributed_object_listener_object_destroyed test
The test failed again recently on the nightly test runner. I could not
reproduce the problem locally, or see a problem in the implementation or test.

To understand the situation better, improved the `assertTrueEventually` utility
method to raise the last error occured and added a context manager
to change the log level for the duration of the test.

Also, improved some parts of the `add_distributed_object_listener` by removing
some unncessary closure definitions.