[3.12.z] - Fix remove_distributed_object_listener test (#281)
In the test, we were creating a map, and then immediately remove
the listener. This works most of the time, but sometimes, event
messages may arrive later, after we got a response from the get_map
request. If we remove the listener after the get_map response, but
before the late event, we will miss the event.

Test is modified to wait for the created event, then removing the
listener, then sleeping for a while and checking that there are no
further events.

Backport of #261
Closes #280