Rewrite scripts
Rewrite the scripts that download the remote controller and run tests
in Python instead of bash or powershell.

The motivation is to reduce duplication. We had to write scripts
for bash and powershell to make it easier to test the client in
different operating systems.

That won't be necessary with a single Python script like this.

Also, added the ability to skip enterprise tests to the test runner
for the users who do not have the HAZELCAST_ENTERPRISE_KEY in
their environment variables.