Set the cluster connection timeout to infinite by default
We decided to set the default cluster connection timeout to
infinite to have a better user experience out-of-the-box
for most of the users/use cases. Note that, this is a breaking
change for the users who rely on the client to shutdown after
some time. Affected users might set the `cluster_connect_timeout`
to a finite value.

Also, updated the default value of the `retry_multiplier` to back off
more after the client tries to connect to the target cluster for
some time.