Do not use deprecated abstract classes from collections (#385)
Abstract classes Sequence and Iterable are meant to be imported
from `collections.abc` instead of `collections` starting from
Python 3.3. Since the minimum supported Python version for the
client is 3.4, we can safely try to import it from the new location
and fallback to the old location if the import fails. It will only
fail for the Python 2.7 in which it is fine to import those abstact
classes from `collections`.