Simplify connection manager (#388)
* Simplify connection manager

Similar to the Java-side fix, we now use the member UUID as the source
of the truth and allow connections to be opened to the same address
more than once. We just make sure that, on the active connections
dictionary, there is just one entry per member via its UUID.

This change simplifies a great deal of things. We now, don't need
a dictionary for pending connections and managing the state of
the connection manager is easier. Also, with this change, we now
make sure that, for the same member, we don't have more than one
active connection in the case of the issue reported on the Java-side
PR.

Also, improved the logging in case of the connection errors.

* call add_backup_listener on unit tests

* fix linter error

* copy active connections while iterating to avoid concurrent modification errors

* fire connected event while not holding the lock