Monkey patch client shutdown on tests for older client versions
Due to a bug on the older client releases, the backward compatibility
tests sometimes fail on the client.shutdown() call. The reasoning
is explained in the https://github.com/hazelcast/hazelcast-python-client/issues/378
and linked PRs.

To solve this problem, we monkey patch the shutdown method, try to
call it, if it fails with an expected exception, cleanup the
leftovers and return without an error.

Shutdown calls generally happen on the teardown methods of the
tests so that we do not leak any clients to other tests. This
change should not cause any problems in actual tests.