Improve Ringbuffer#read_many
Before this, read_many was returning an immutable lazy list of
items read. However, there were more information sent by the server
which we were omitting.

Now, read_many returns an instance of ReadResult, which the user
can get the read count, size, next sequence to read from and
item sequences. Note that, this is not a breaking change. ReadResult
inherits from the ImmutableLazyDataList and hence is a list. So, the
users of this API can use their code as it is without any change when
they upgrade to 4.1. New users might benefit from the new APIs provided.

Also added a filter parameter to read_many, refactored the implementation
and updated some docstrings.