[API-339] Implement ReliableTopic (#395)
* Initial ReliableTopic implementation: #201

* Client does not have _listener, but instead listener as the reference to the listener service

* Remove leftovers and useless comments

* Correct documentation: ReliableTopicListener -> reliable_topic_listener

* Prevent deadlock in error handler

* Add new tests: test_client_not_active_error, test_distributed_object_destroyed_error, test_distributed_object_destroyed, test_stale

* Implement ReliableTopic

Implemented ReliableTopic on top of the Ringbuffer. Like the
interface provided in the Java-side, we are offering methods
to add/remove listeners and publish messages in batches or
one by one. The provided API is non-blocking just like the
other parts of the client.

Documentation and code samples will be sent in a separate PR.

* construct memberinfo only if the publisher_address is non-none

* add durable subscription test

Co-authored-by: Burak Sezer <burak.sezer.developer@gmail.com>