Backward compatibility tests related fixes
Fixed some of the tests that would fail when run against the different
Hazelcast IMDG versions.

There are two of them.

- 4.2+ servers do not collect per-entry stats by default, so when
the client is running on such a cluster, get_entry_view will return
unset values for some stats.

- For one of the SSL tests, we start the member with a CA-signed
certificate and on the client side, we just enable the SSL. For this
to work, we need to regularly update the certificate(since it expires).
Also, old server JARs contain the expired version of the certificate.
To workaround this, instead of relying on the server test JAR to contain
such certificates, we will put the certificate on the client repo and
maintain an up-to-date version there.