Fix SOAK test script (#400)
The Python script we are using were trying to import something
from the test package. That was not good because inside the
`__init__` file of the test package, we define a global logger
config that we want to use in tests.

But, in the SOAK test, we are using some other logger config.

To make the logger output work, import is removed and the functionnality
is directly copy-pasted, since it is something really small.