Add SQL code sample
Added a code sample that demonstrates different ways to interact
with the SQL service and the returned SqlResult objects.

Also, removed the explanation of the folders in the README because
it was out of sync with the current folder structure and was
not offering lots of value. It was mainly just repeating the
folder names.