Show the context manager support for transactions in code samples and documentation (#410)
We do have a context manager support for transactions but we were
not showing it in our code samples or documentation. As a result
of that, users were unaware of this feature. We even had a call
with one user who implemented a simple context manager for
Hazelcast transactions himself.

So, with this PR, we are now showing that we have support for it
in our code samples and documentation.