Fire membership events when HotRestart is enabled
With this change, we now clear the memberlist when the cluster changes,
instead of just clearing the memberlist version so that, we can
fire events when a member with HotRestart enabled restarts with
the same UUID. Added a test that verifies this behavior.

Also, removed an unnecessary usage of the `clear_member_list_version`
method.