Fix `test_lifecycle_listener_receives_disconnected_event`
This test failed in Github Actions runners. The reason of the failure
is that, we try to check the received events right after shutting down
the member. It might be the case that, the member is shutdown, but
it is not visible to the client yet.

To tackle that, we now use `assertTrueEventually` to verify that
we receive the event.