Check the enterprise key properly in the test runner
In the test runner, we just check if there is an entry named
`HAZELCAST_ENTERPRISE_KEY` in the environment variables. For our
codecov runner, for PRs, it is in fact there but set to an empty value.

Hence, it was passing the check mistakenly. Now, we don't start
the enterprise tests, if there is an empty enterprise key.