[API-537] Add support for the binary metrics (#441)
* Add support for the binary metrics

Implemented the mechanism to compress the metrics we collect in a
way that is compatible with the server-side.

With this change, the latest Management Center versions are able to
show statistics regarding the Python client.

For backward compatibility, we also continue sending metrics in
a comma separated key-value format (client attributes).

In this PR, I also significantly cleaned up the statistics service.
In the old code, we were trying to reuse the results of some
calls. In the benchmarks I did, I saw that it didn't really improve
much. (New approach approx. 0.5 ms, old approach approx. 0.4 ms per
statistics collection) This really small difference does not justify
the need for complexity so I went with the simpler approach.

Also, aligned the log levels with the Java client for the various
warnings.

Also, bumped the version of psutil to 5.8.0.

* address review comments