Improve test runner
When there are some scripts in the main directory, the test runner
might get stuck on executing it and fail to start tests.

To improve the local development experience, I added the directory
of the tests to the test runner so that it only tries to run
tests from there.

Also, added some common virtual environment directories to the .gitignore.