v5 SQL: Initial PR
This is an initial PR for the v5 SQL. It consists minimal changes
required to test the v5 SQL. Follow up PRs will come after that.

Here are the list of changes made:

- Set client version to 5.0
- Update remote controller scripts so that it uses hazelcast+hazelcast-sql
JARs instead of hazelcast-all, since hazelcast-all is not produced anymore.
Also, update some documentation about it.
- Use ints to represent years in DATE column type
- Use built-in types for DATE, TIME, TIMESTAMP, TIMESTAMP WITH TIMEZONE, and DECIMAL
which are datetime.date, datetime.time, datetime.datetime,
datetime.datetime(with non-None tzinfo), and decimal.Decimal.
- Update SQL tests and add some Jet tests
- Fix distributed objects tests by filtering internal objects out.