v5 SQL: Remove unused load balancer methods
These methods were specially added for 4.2 SQL, but they are unused
in v5. Hence, we are removing the unused methods and reverting the
changes we made around the load balancer implementations and tests.

Note that, it is perfectly safe to remove these methods in Python.
If there were users that implemented these methods, they can use
their code without changing anything.