Fix Github Actions runners (#465)
In the nightly tests, there were some failures.

- In Windows runners, `test_await_latch_with_timeout` was failing
randomly. My assumption is that, due to low time resolution, sometimes
we were getting equal values instead of larger values. I changed
the condition from `>` to `>=`. Also, I changed the failure message
so that, the next time it fails we will see the exact values.

- Removed `python -m pip install --upgrade pip` statements as the Python
runners already have the latest pip available to them. If we try to
upgrade pip to the latest version, for some older Python versions
it fails because those Python versions reached their EOL and pip
no longer tries to maintain compatibility for them.

- Modified the nightly runner so that, when the os is `ubuntu-latest`
and the Python version is `3.4`, we will use `ubuntu-18.04`, as there
is no Python 3.4 binary available in `ubuntu-latest`. Hopefully, we
will drop the support for ancient Python versions, but till that
day, this workaround should work.