Implement external smart client discovery
This PR implements the client-side support for the external smart
client discovery. With this PR, the Python client will be able
to use public addresses reported by the cluster members even
if the client and the cluster are in different networks.

To use it, one has to enable `use_public_addresses` option by setting
it to `True`, and passing at least one public address of the cluster
members to the `cluster_members` config. Client will initiate the
cluster connection using the provided public address, and will
discover the rest of the cluster using the public addresses reported
by them.