Increase timeout used in `test_await_latch_with_timeout` (#469)
* Increase timeout used in `test_await_latch_with_timeout`

Increased the timeout from 0.1 second to 1 second. Also, in Windows
due to low time resolution, we lowered the compared value slightly.

With increased timeout, comparing to a slightly lower value should
be OK since it shows that we are in fact waiting for the timeout.

Note that, the miscalculation comes from the fact that, the calculation
for the time passed in the Python side is precise to +-16ms. See
https://bugs.python.org/issue44328 and the Windows documentation for
the methods used by the CPython for more information.

* correctly print the message