Provide suggestions for the `HazelcastSqlError`
Recently, an enhancement is implemented on the server-side to provide
suggestions to execute in case of SqlErrors. The main use case is to
provide the `CREATE MAPPING` query suggestions in case there exists
a map and there are some entries on it.

The documentation of this feature will be provided in the upcoming
SQL documentation update PR.