[API-704] v5 SQL: Update documentation (#461)
* v5 SQL: Update documentation

Updates the SQL documentation with the changes done in the reference
manual. This change mainly reorganizes the documentation and adds some
notes about the new capabilities. For the details, we directed users
to the reference manual so that we don't have to duplicate most parts
of the reference manual which can change frequently.

* address review comments

* upgrade sphinx and sphinx-rtd-theme dependencies