Fix re-registration logic of the ClusterViewListener
There was a test failure due to StackOverflow error while trying
to re-register the cluster view listener.

The chain of events that caused the overflow are following:

- client.shutdown() is called
- we have a couple of connections
- while closing the connection manager, we close connections one-by-one
- we close the connection that the cluster view listener is registered to
- after closing this, we trigger a re-registration attempt and try to invoke
the listener registration in one of the remaining connections
- since the client is shutting down, the invocations are failed immediately
- while trying the add a callback to the invocation future, since it
is already done, we call the callback immediately
- that causes an infinite chain of registration attempts and leads the
StackOverflow error

There are two solutions to this problem. While either one of them would
resolve the issue, I decided to apply both to make this re-registration
logic more solid.

- The first solution is to not attempt to register a listener if the client
is about to shut down.

- The second solution is to schedule the next re-registration attempt in
the next iteration of the event loop, so that we would allow the thread
that calls shutdown to complete its work.