Deserialize `SqlRow`s lazily (#472)
This is a required change to be able to use parts of the row,
even if we cannot deserialize some column values.

Also, lazy deserialization is a plus on its own.

To accomplish this, similar to the Java client, I have changed the
code so that, we try to deserialize column values when we
call `get_object` or `get_object_with_index`. Note that,
just like the Java client

```
row.get_object("smth")
row.get_object("smth")
```

will result in two deserializations.
We could easily solve this with an extra object per row (something
like LazyDeserializable class with two fields: serialized and deserialized
where once we deserialize it, it assigns the result to deserialized and
return it for the subsequent calls). But, I didn't want to go that way due
to extra object cost.