Make `SqlRow` support `[]` operator
This is mainly an enhancement for easier access for columns in the row.
Now, the `SqlRow` implements the `__getitem__` magic method.

Implementation of the method simply performs a type check and calls
`get_object_with_index` if the type of the passed parameter is an
integer type. For any other type, it calls `get_object`. Hence,
it will throw for any other type than `int` or `str`.

Also, put a piece of information about this new way to the documentation
& code samples.