Wait for the initial response in the SQL queries
Up until now, we were returning the `SqlResult` immediately after
initiating the query to the server, without waiting for the first
response.

The server might not be able to handle some consequences of such
an API. To align the behavior with the Java implementation and
make it more intuitive, we now wait for the initial response
in the `execute` and `execute_statement` methods.

With that change, the return types are changed from `SqlResult` to
`Future[SqlResult]`. Also, I got rid of the Futures in the return types
of the methods over the `SqlResult` except for the `close`, as the
`close` might in fact go to the server.

With this change, I cleaned up the implementation so that it reflects
the new behavior better.