Update SQL documentation according to recent changes
This PR updates the SQL documentation at the client-side with
the recent changes done to the reference manual.

We mostly follow a similar structure, but we go into more detail
for Portable and JSON queries in these documents as these are
the most intuitive ways to run SQL queries from the Python client.

Note that, the code samples in this PR are written according to
recent PRs, some of which not merged yet.