Get rid of the `execute_statement` API for SQL
We were providing two APIs two execute SQL queries.

- `execute` with a string query and parameters
- `execute_statement` with `SqlStatement` parameter two have more
control over the query execution.

We decided to get rid of the `execute_statement` and `SqlStatement`
and provide ability to pass keyword arguments to `execute` to customize
the behavior of the queries. With that change, the properties of
`SqlStatement` is now keyword arguments.

Also, upgraded the server version to 5.0 as 5.0 is released.