Reduce test duration for the integration tests (#483)
There were some tests that are taking too much time to run. Most
of the time, we either create too many members or sleep too much
while waiting on some assertions.

This is a PR to tackle with that. I went over the long running
tests and either shorten the duration of the waits (of course without
breaking the test) or reduced the number of members we create without
changing the behavior of the tests.