Improve SOAK test script
In the last SOAK test run, we saw a false positive being produced
during the shutdown. The following scenario happened in the run:

- Assume that the test deadline is 16:02:15
- Assume that we last displayed operation counts at 16:02:14
- After displaying the operation counts, the displayer thread
goes to sleep for 10 seconds.
- After a moments later the threads that run operations see that
the deadline is reached and stopped working.
- Some threads stopped even earlier than the first operation
after the displayer thread reset their operation count.
- When the displayer thread wakes up after 10 seconds, it will
see that some threads did no work, but they need not to, because
the deadline was reached.
- The displayer thread falsely assumes that such threads are hanged.

To solve this problem, we delegate the task of checking the deadline
just to the observer thread. It will set a boolean flag if the deadline
is reached, and the test runners will stop. When the observer wakes
up from sleep and sees that the deadline is reached, it will
not check the operation counts, to deal with the problem mentioned
above.

Also, did some refactorings to scripts, readme, and the test runner.