Update GA workflow files for the 4.2.z branch
While sending a PR to the 4.2.z branch, I saw that it uses the GA
workflows defined in that branch, so I am updating the test runner
and jira trigger workflows so that, the test runner creates
jobs that would pass the merge checks.

The only difference between this test runner and the one in the
master is that, this one does not check broken links in the readme.