Improve the client version filtering in `ClientTcpMetricsTest` (#504)
The features tested in `ClientTcpMetricsTest` are backported into
`4.2.2` and `5.0.1` as well. So, we want to include this test
in their backward compatibility tests.

To do so, the filter is changed so that it will skip tests when
the client version is older than `4.2.2` or equal to `5.0`, which
skips the client versions that do not have such feature.