Drop support for Python 2.7, 3.4, and 3.5
This PR drops support for Python 2.7, 3.4, and 3.5 and adds a matrix
item to test the client code on Python 3.10.

Those aforementioned versions are removed from/added to the corresponding
workflow matrix.

I also modified the PR runner so that it runs tests only on the minimum
Python version (3.6) we support, as I believe we should take it into
consideration while making changes.

Also, I have removed now unnecessary Python version checks from the
requirements-*.txt files and modified the setup.py to reflect which
Python versions we now support.